Some minor refinements
