+++ src/main/java/io/quarkus/bot/AffectMilestone.java
@@ -23,6 +23,7 @@ import io.quarkus.bot.config.QuarkusGitHubBotConfigFile;
 import io.quarkus.bot.util.GHIssues;
 import io.quarkus.bot.util.IssueExtractor;
 import io.quarkus.bot.util.Labels;
+import io.quarkus.bot.util.Strings;
 import io.smallrye.graphql.client.dynamic.api.DynamicGraphQLClient;
 
 class AffectMilestone {
@@ -104,8 +105,9 @@ class AffectMilestone {
         }
         if (!comment.isEmpty()) {
             comment = "Milestone is already set for some of the items:\n\n" + comment;
-            comment += "\nWe haven't automatically updated the milestones for these items.\n";
-            comment += "\n> This message is automatically generated by a bot.";
+            comment += "\nWe haven't automatically updated the milestones for these items.";
+
+            comment = Strings.commentByBot(comment);
 
             if (!quarkusBotConfig.isDryRun()) {
                 pullRequest.comment(comment);
