+++ src/main/java/io/quarkus/bot/CheckPullRequestEditorialRules.java
@@ -19,6 +19,7 @@ import io.quarkiverse.githubapp.event.PullRequest;
 import io.quarkus.bot.config.Feature;
 import io.quarkus.bot.config.QuarkusGitHubBotConfig;
 import io.quarkus.bot.config.QuarkusGitHubBotConfigFile;
+import io.quarkus.bot.util.Strings;
 
 class CheckPullRequestEditorialRules {
 
@@ -57,10 +58,9 @@ class CheckPullRequestEditorialRules {
         for (String errorMessage : errorMessages) {
             comment.append("- ").append(errorMessage).append("\n");
         }
-        comment.append("\n> This message is automatically generated by a bot.");
 
         if (!quarkusBotConfig.isDryRun()) {
-            pullRequest.comment(comment.toString());
+            pullRequest.comment(Strings.commentByBot(comment.toString()));
         } else {
             LOG.info("Pull request #" + pullRequest.getNumber() + " - Add comment " + comment.toString());
         }
