+++ src/main/java/io/quarkus/bot/util/Strings.java
@@ -10,6 +10,15 @@ public class Strings {
         return string == null || string.trim().isEmpty();
     }
 
+    public static String commentByBot(String originalComment) {
+        StringBuilder sb = new StringBuilder(originalComment);
+        if (!originalComment.endsWith("\n")) {
+            sb.append("\n");
+        }
+        sb.append("\n> This message is automatically generated by a bot.");
+        return sb.toString();
+    }
+
     private Strings() {
     }
 }
