+++ src/test/java/io/quarkus/bot/it/CheckIssueEditorialRulesTest.java
@@ -28,47 +28,4 @@ public class CheckIssueEditorialRulesTest {
                 });
 
     }
-
-    @Test
-    void testLabelBackportWarningConfirmation() throws IOException {
-        String warningMsg = String.format(CheckIssueEditorialRules.LABEL_BACKPORT_WARNING, "triage/backport-whatever");
-        StringBuilder expectedComment = new StringBuilder("@test-github-user " + warningMsg);
-        expectedComment.append("\n> This message is automatically generated by a bot.");
-
-        given().github(mocks -> mocks.configFile("quarkus-github-bot.yml").fromString("features: [ ALL ]\n"))
-                .when().payloadFromString(getSampleIssueLabelTriageBackportPayload())
-                .event(GHEvent.ISSUES)
-                .then().github(mocks -> {
-                    verify(mocks.issue(1234567890))
-                            .comment(expectedComment.toString());
-                    verifyNoMoreInteractions(mocks.ghObjects());
-                });
-
-    }
-
-    private static String getSampleIssueLabelTriageBackportPayload() {
-        return """
-                {
-                    "action": "labeled",
-                    "issue": {
-                      "id": 1234567890,
-                      "number": 123,
-                      "labels": [
-                        {
-                          "name": "triage/backport-whatever"
-                        }
-                      ]
-                    },
-                    "label": {
-                      "name": "triage/backport-whatever"
-                    },
-                    "repository": {
-
-                    },
-                    "sender": {
-                      "login": "test-github-user"
-                    }
-                  }
-                    """;
-    }
 }
